identity: move identity restorer from daemon/cmd to pkg/identity/restoration
This comit moves the extracted identity restorer from package `daemon/cmd`
to `pkg/identity/restoration`.
